requirements.txt
numpy==2.4.2
pandas==2.3.3
scikit-learn==1.8.0
torch==2.10.0+cpu
torchvision==0.25.0+cpu
pillow==12.1.1